refactor(auth): add index barrel file and simplify imports
