+++ frontend/src/features/auth/index.js
@@ -1,0 +1,3 @@
+export * from './authSlice';
+export * from './authActions';
+export * from './hooks/useAuth';
